requirements.txt
svgwrite
numpy
cairosvg
potracer